+++ NewZhongYan/NewZhongYan/SKGtaskDetailController.m
@@ -1673,7 +1673,7 @@
 -(void)getPhrase:(UIButton*)sender
 {
     SKCommonLanguageController *lc = [[SKCommonLanguageController alloc] init];
-    lc.textViewKey=[NSString stringWithFormat:@"%d",sender.tag/10000];
+    lc.textViewKey=[NSString stringWithFormat:@"%d",(int)(sender.tag/10000)];
     HPGrowingTextView *textV=(HPGrowingTextView *)[self.view viewWithTag:((UIButton *)sender).tag/10000];
     lc.textViewText=textV.text;
     currentTextViewindex = [saveTextViewArray indexOfObject:textV];
